created dashboard foundation. Organized files to make navigation a little easier
