+++ ClientApp/src/components/Donor/DonorDashboard.js
@@ -1,0 +1,20 @@
+﻿import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
+
+export class DonorDashboard extends Component{
+    render() {
+        return (
+            <div>
+                Hello logged-in donor!
+                <div class="DonorDashboardNavbar">
+                    <ul>
+                        <li>Questionaire</li>
+                        <li>Schedule An Appointment</li>
+                        <li>Locate Blood Donation Sites</li>
+                        <li><Link to="/edit-donor-info">Edit Account Information</Link></li>
+                    </ul>
+                </div>
+            </div>
+        );
+    }
+}
