+++ ClientApp/src/components/Donor/DonorLogin.js
@@ -1,23 +1,24 @@
 ﻿import React, { createContext, Component, useContext } from 'react';
 import { useNavigate } from 'react-router-dom'
-import { AuthContext } from '../context/AuthContext'
+import { AuthContext } from '../../context/AuthContext'
 
 // the only reason I'm keeping this stuff class-based instead of switching it over to a function is because we're too tight on time right now 
 function DonorLoginFunction(props) {
     const navigate = useNavigate();
     const { login } = useContext(AuthContext)
-    return <DonorLogin navigate={navigate}/>
+    return <DonorLogin navigate={navigate} login={login} />
 }
 class DonorLogin extends Component {
     static displayName = DonorLogin.name;
     static contextType = AuthContext;
+    
     constructor(props) {
         super(props);
         this.state = {
             username: '',
             password: ''
         }
-
+        this.login = props.login;
         this.handleChange = this.handleChange.bind(this)
         this.handleLoginAttempt = this.handleLoginAttempt.bind(this)
     };
@@ -36,8 +37,8 @@ class DonorLogin extends Component {
 
     handleLoginAttempt = (event) => {
         event.preventDefault();
-        const auth = this.context;
-        this.context.login(this.state.username, this.state.password)
+        debugger;
+        this.login(this.state.username, this.state.password)
             .then(() => {
                 console.log("success")
             })
