+++ ClientApp/src/components/Donor/EditDonorInfo.js
@@ -1,0 +1,12 @@
+﻿import React, { Component } from 'react';
+
+export class EditDonorInfo extends Component {
+    render() {
+        return (
+            <div>
+                This is where donors go to edit their personal information.
+                
+            </div>
+        );
+    }
+}
