+++ ClientApp/src/components/Donor/RegisterDonor.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import states from './StatesList';
+import states from '../StatesList';
 
 export class RegisterDonor extends Component {
     static displayName = RegisterDonor.name;
