--- ClientApp/src/components/DonorDashboard.js
@@ -1,11 +1,0 @@
-﻿import React, { Component } from 'react';
-
-export class DonorDashboard extends Component{
-    render() {
-        return (
-            <div>
-                Hello logged-in donor!
-            </div>
-        );
-    }
-}
