+++ ClientApp/src/components/Employee/EmployeeDashboard.js
@@ -1,0 +1,19 @@
+﻿import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
+
+export class EmployeeDashboard extends Component {
+    render() {
+        return (
+            <div>
+                Hello logged-in Employee!
+                <div class="DonorDashboardNavbar">
+                    <ul>
+                        <li>List of Donors</li>
+                        <li>Blood Stock</li>
+                        <li><Link to="/edit-employee-info">Edit Account Information</Link></li>
+                    </ul>
+                </div>
+            </div>
+        );
+    }
+}
